feat(properties): update property schema and validation
- Changed landSize and floorArea types from decimal to integer in property schema
- Added validation constraints for landSize (max 1,000,000 sqm) and floorArea (max 100,000 sqm)
- Updated search schema to use integer types for land and floor area filters
- Added new migration (0004_nifty_kitty_pryde) to reflect schema changes
- Updated property creation and update schemas with proper integer validation
- Added meaningful validation messages for size constraints

This change ensures consistent data types between the application and database layers, improves data integrity by using appropriate integer types for size measurements, and provides better validation feedback to users.

+++ src/schemas/propertySearch.schema.ts
@@ -36,10 +36,10 @@ export const propertySearchSchema = z
     maxBathrooms: z.coerce.number().int().nonnegative().optional(),
     minParkingSpaces: z.coerce.number().int().nonnegative().optional(),
     maxParkingSpaces: z.coerce.number().int().nonnegative().optional(),
-    minLandSize: z.coerce.number().nonnegative().optional(),
-    maxLandSize: z.coerce.number().nonnegative().optional(),
-    minFloorArea: z.coerce.number().nonnegative().optional(),
-    maxFloorArea: z.coerce.number().nonnegative().optional(),
+    minLandSize: z.coerce.number().int().nonnegative().optional(),
+    maxLandSize: z.coerce.number().int().nonnegative().optional(),
+    minFloorArea: z.coerce.number().int().nonnegative().optional(),
+    maxFloorArea: z.coerce.number().int().nonnegative().optional(),
 
     // Property features
     hasAirConditioning: z.coerce.boolean().optional(),
